Keep track of players
Will now keep track of connected players and their statuses as well as
which players are in the channels that the bot is in. The bot can now
also get the player's profile data from the JSON endpoints.

This commit adds state to the bot which makes it more complex but will
allow for more interesting commands.