Fix searching for just first word of player name
Previously if we had two players "a bread" and "apple", then player "a
bread" was not possible to be selected because the bot only checked the
first word "a" which is also the letter that "apple" starts with. With
this change, the bot joins "a" and "bread" with space before searching.